1st round of tests
These classes do not meet all proper verification instructions.   A valid email address will consist of the following: a local part that must begin with a letter, followed by a sequence of numbers and/or letters. No punctuation, special symbols, etc. Then a single @ symbol. Then a domain name which has the same rules as the local part above but it may include "."'s. The final "." must be followed by "edu", "com", "biz", or "gov", and those should be the final characters in the email address. The local part and domain name must have at least one character. For SS class, social security numbers with all zeros in any digit group (000-##-####, ###-00-####, ###-##-0000) are not allowed. B. Numbers with 666 or 900-999 in the first digit group are invalid. C. The number 078-05-1120 is invalid.